requirements.txt
discord.py
python-dotenv
transformers
torch
GitPython
